created general interfaces folder for public types and interfaces
