+++ waterloop-site/src/components/Profiles/components/ProfileSection.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { ProfileType, PSectionProps, PSectionState } from '../interfaces'
+import ProfileType from '../../../interfaces'
+import { PSectionProps, PSectionState } from '../interfaces'
 import { SubProfile, LeadProfile, ExpandedProfile } from '../../Profiles'
 
 // Styled components for ProfileSection
