+++ waterloop-site/src/components/Profiles/interfaces.tsx
@@ -1,15 +1,7 @@
+import ProfileType from '../../interfaces'
+
 export type ContactsType = {
  contacts: Array<{icon: string, url: string}>
-}
-
-export type ProfileType = {
-  name: string,
-  position: string,
-  programInfo: string,
-  portrait: string,
-  teams: Array<string>,
-  bio: string,
-  contacts: Array<{icon: string, url: string}>
 }
 
 export interface PSectionProps {
