+++ waterloop-site/src/interfaces/index.tsx
@@ -1,0 +1,3 @@
+import ProfileType from './ProfileType'
+
+export default ProfileType
